pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agora-protocol"
version = "0.1.0"
description = "On-chain debate-or-vote arbitration for multi-agent LLM systems"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.3.0",
    "langchain-google-vertexai>=2.0",
    "langchain-core>=0.3.0",
    "google-cloud-aiplatform>=1.70",
    "pydantic>=2.0",
    "structlog>=24.0",
    "numpy>=1.26",
    "merkletools>=1.0.3; python_version < '3.11'",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
include = ["agora*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
